Enable the unique_id for GPU devices

+++ client/src/rpc_client.rs
@@ -13,7 +13,7 @@ pub trait RpcClient {
 
     async fn check_server(&self) -> Result<(), Error>;
 
-    async fn list_allocations(&self) -> Result<Vec<(usize, u64)>, Error>;
+    async fn list_allocations(&self) -> Result<Vec<(u64, u64)>, Error>;
 
     async fn release(&self, client: ClientToken) -> Result<(), Error>;
